huggingface/requirements.txt
gradio
ultralytics
torch
numpy
pillow
opencv-python-headless